Update PDM config for version numbering.
The actual fundamental reason that many builds are failing locally right now
is that the tags we have use v0.1.2.devX where x is a number greater than 0,
and setuptools_scm doesn't support this. So we need to use dev0 if anything.
So in this change I added a new tag (which we really shouldn't be doing on a
branch!) just to reset the most recent tag in history so we don't encounter
this build problem anymore.

I also added some more config for the kinds of version numbers used and how
they are incremented and some more config for ruff linting (even though we're
not enforcing it yet).

Signed-off-by: Richard Bair <rbair23@users.noreply.github.com>